mejoras en el responsive y correcciones de errores

+++ Controlador/js/EnviarPool.js
@@ -1,4 +1,5 @@
 var formularioEnvioPool = document.getElementById("formularioPool");
+var mensajeError = document.getElementById("mensajeError")
 
 formularioEnvioPool.addEventListener("submit", function (e) {
   e.preventDefault();
@@ -8,7 +9,24 @@ formularioEnvioPool.addEventListener("submit", function (e) {
 function EnvioPool() {
   var datos = new FormData(formularioEnvioPool);
   var nombreTorneo = datos.get("nombreTorneo");
-  localStorage.setItem('nombreTorneo', nombreTorneo); 
-  window.location.href="../Pool/CreadorPools.html"
+  alert(nombreTorneo)
+  var existe=datosingresados(nombreTorneo);
+  if(existe){
+    var myModal = new bootstrap.Modal(document.getElementById("exampleModal"));
+    myModal.show();
+    mensajeError.innerHTML ="Ingrese un torneo";
+  }
+  else{
+    localStorage.setItem('nombreTorneo', nombreTorneo); 
+    window.location.href="../Pool/CreadorPools.html"
+  }
+}
 
+function datosingresados(dato){
+  if(dato.includes("Ingrese")){
+       return true
+  }
+  else{
+    return false
+  }
 }
