+++ Controlador/js/EnvioElegirParticipante.js
@@ -7,7 +7,22 @@ formularioTanteador.addEventListener("submit", function (e) {
 function EnvioTanteador() {
   var datos = new FormData(formularioTanteador);
   var nombreTorneo = datos.get("nombreTorneo");
-  alert("nombre del torneo:" + nombreTorneo);
-  localStorage.setItem('nombreTorneo', nombreTorneo); 
-  window.location.href="../Participante/ElegirParticipante.html"
+  var existe=datosIngresados(nombreTorneo)
+  if(existe){
+    alert ("ingrese un torneo para continuar");
+  }
+  else{
+    alert("nombre del torneo:" + nombreTorneo);
+    localStorage.setItem('nombreTorneo', nombreTorneo); 
+    window.location.href="../Participante/ElegirParticipante.html"
+  }
+}
+
+function datosIngresados(dato){
+ if(dato.includes("Ingrese")){
+  return true
+ }
+ else{
+  false
+ }
 }
